apply review - move  ‘getFormattedExperiment’ and ‘getFormmattedExperimentList’ helpers to a new helper ‘formatDatasetAction’ and adjust the imports
