+++ src/components/SearchResults/AddPageToDatasetButton.js
@@ -1,5 +1,5 @@
 import { useResponsive } from 'hooks/useResponsive'
-import getFormattedExperimentList from 'helpers/getFormattedExperimentList'
+import { getFormattedExperimentList } from 'helpers/formatDatasetAction'
 import { DatasetActionButton } from 'components/shared/DatasetActionButton'
 
 export const AddPageToDatasetButton = ({ dataToAdd }) => {
