+++ src/components/shared/SearchCard/SearchCardAction/SearchCardAction.js
@@ -2,7 +2,7 @@ import { memo } from 'react'
 import { Box } from 'grommet'
 import { useDatasetManager } from 'hooks/useDatasetManager'
 import { useResponsive } from 'hooks/useResponsive'
-import getFormattedExperiment from 'helpers/getFormattedExperiment'
+import { getFormattedExperiment } from 'helpers/formatDatasetAction'
 import { DatasetActionButton } from 'components/shared/DatasetActionButton'
 import { DownloadNowButton } from './DownloadNowButton'
 import { ProcessingDatasetPill } from './ProcessingDatasetPill'
